+++ eslint.config.mjs
@@ -1,6 +1,8 @@
 import { generateEslintConfig } from '@companion-module/tools/eslint/config.mjs'
 
-const baseConfig = await generateEslintConfig({})
+const baseConfig = await generateEslintConfig({
+	enableTypescript: true,
+})
 
 const customConfig = [
 	...baseConfig,
