+++ package.json
@@ -1,15 +1,17 @@
 {
 	"name": "generic-swp08",
-	"version": "2.0.11",
+	"version": "3.0.0",
 	"main": "src/index.js",
 	"type": "module",
 	"scripts": {
-		"prepare": "husky",
-		"test": "echo \"Error: no test specified\" && exit 1",
-		"format": "prettier --w .",
-		"package": "companion-module-build",
+		"postinstall": "husky",
+		"format": "prettier -w .",
+		"package": "run build && companion-module-build",
+		"build": "rimraf dist && run build:main",
+		"build:main": "tsc -p tsconfig.build.json",
+		"dev": "tsc -p tsconfig.build.json --watch",
 		"lint:raw": "eslint",
-		"lint": "yarn lint:raw ."
+		"lint": "run lint:raw ."
 	},
 	"license": "MIT",
 	"repository": {
@@ -24,13 +26,17 @@
 	"devDependencies": {
 		"@companion-module/tools": "^2.7.1",
 		"@types/lodash": "^4.17.24",
+		"@types/node": "^22.19.17",
 		"eslint": "^9.39.4",
 		"husky": "^9.1.7",
 		"lint-staged": "^16.4.0",
-		"prettier": "^3.8.3"
+		"prettier": "^3.8.3",
+		"rimraf": "^6.1.3",
+		"typescript": "^6.0.3",
+		"typescript-eslint": "^8.58.2"
 	},
 	"engines": {
-		"node": "^22.13"
+		"node": "^22.17"
 	},
 	"prettier": "@companion-module/tools/.prettierrc.json",
 	"packageManager": "yarn@4.14.1",
