--- src/upgrades.js
@@ -1,6 +1,0 @@
-export const UpgradeScripts = [
-	/*
-	 * Place your upgrade scripts here
-	 * Remember that once it has been added it cannot be removed!
-	 */
-]
