+++ src/upgrades.ts
@@ -1,0 +1,8 @@
+import type { CompanionStaticUpgradeScript } from '@companion-module/base'
+import type { SwP08Config } from './config.js'
+export const UpgradeScripts: CompanionStaticUpgradeScript<SwP08Config>[] = [
+	/*
+	 * Place your upgrade scripts here
+	 * Remember that once it has been added it cannot be removed!
+	 */
+]
