+++ tsconfig.build.json
@@ -1,0 +1,15 @@
+{
+	"extends": "@companion-module/tools/tsconfig/node22/recommended-esm.json",
+	"include": ["src/**/*.ts"],
+	"exclude": ["node_modules/**", "src/**/*spec.ts", "src/**/__tests__/*", "src/**/__mocks__/*"],
+	"compilerOptions": {
+		"outDir": "./dist",
+		"rootDir": "./src",
+		"paths": {
+			"*": ["./node_modules/*"]
+		},
+		"module": "nodenext",
+		"moduleResolution": "nodenext",
+		"alwaysStrict": true
+	}
+}
